+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version 3.1.2 *(In Development)*
 * Fix: `MenuItem.getMenuInfo()` was throwing runtime exception. Will now just
   return `null`.
 * Fix: Dragging over a `WebView` contained in a `ViewPager` would not register.
+* Fix: Inflation of context menu incorrectly being handled by the custom menu
+  inflater for the library.
 
 
 Version 3.1.1 *(2011-07-31)*